requirements.txt
numpy==1.19.5
scikit-image==0.17.2
opencv-python==4.5.1.48
tensorflow==2.4.1
tensorflow_datasets==4.2.0
tensorflow_addons==0.12.1
-i https://google-coral.github.io/py-repo/ tflite_runtime
matplotlib==3.3.4
Pillow==8.1.1
ipython==7.16.1